# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dehasoft E-Ticaret Çözümü",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/card/get": {
            "get": {
                "tags": [
                    "Kullanıcı Kart"
                ],
                "summary": "Kartları getir",
                "description": "Kullanıcının kayıtlı kartlarını listeler.",
                "operationId": "userCardGet",
                "requestBody": {
                    "description": "Kartları listeler.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": []
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kullanıcı kartları başarı ile sorgulandı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kullanıcı kartları başarı ile sorgulandı"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/user/card/create": {
            "post": {
                "tags": [
                    "Kullanıcı Kart"
                ],
                "summary": "Kart Oluştur",
                "description": "Kullanıcı kart kaydeder.",
                "operationId": "userCardCreate",
                "requestBody": {
                    "description": "Kartları oluşturur.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "card_expire",
                                    "card_cvv",
                                    "card_name"
                                ],
                                "properties": {
                                    "card_number": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "card_expire": {
                                        "type": "string",
                                        "example": "12/20"
                                    },
                                    "card_cvv": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "card_name": {
                                        "type": "string",
                                        "example": "Banka Kartım"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kullanıcı kartı başarı ile oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kullanıcı kartı başarı ile oluşturuldu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/user/card/delete": {
            "post": {
                "tags": [
                    "Kullanıcı Kart"
                ],
                "summary": "Kart Sil",
                "description": "Kullanıcının kayıtlı kartı siler.",
                "operationId": "userCardDelete",
                "requestBody": {
                    "description": "Kartı siler.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_id"
                                ],
                                "properties": {
                                    "card_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kullanıcı kartı başarı ile güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kullanıcı kartı başarı ile güncellendi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
        "/api/user/card/update": {
            "post": {
                "tags": [
                    "Kullanıcı Kart"
                ],
                "summary": "Kart Güncelle",
                "description": "Kullanıcının kayıtlı kartı günceller.",
                "operationId": "userCardUpdate",
                "requestBody": {
                    "description": "Kartları günceller.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": []
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kullanıcı kartı başarı ile güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kullanıcı kartı başarı ile güncellendi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "deha_token": {
                "type": "apiKey",
                "description": "Login işleminden dönen tokenType ve tokeni sırasıyla boşluk bırakarak giriniz.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}